requirements.txt
python-telegram-bot[ext]==21.0.1
shazamio==0.5.1
pytz==2024.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
yt-dlp==2024.05.27
wit==6.5.0